requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
hupper==1.12.1
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
pdfkit==1.0.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
urllib3==2.2.1
